Move aMatrixRoom to its own file, since it's used by several test classes.
